a lot of changes
int -> qreal; set width & height property only under change; code refactor; v8 isolate npe try fix

+++ doric-Qt/doric/shader/DoricVLayoutNode.cpp
@@ -22,7 +22,7 @@ QQuickItem *DoricVLayoutNode::build() {
 void DoricVLayoutNode::blend(QQuickItem *view, QString name, QJsonValue prop) {
   QQuickItem *container = view;
   if (name == "space") {
-    getLayouts()->setSpacing(prop.toInt());
+    getLayouts()->setSpacing(prop.toDouble());
   } else if (name == "gravity") {
     getLayouts()->setGravity(prop.toInt());
   } else {
